chore: decouple composite input generator from fuzzer
