+++ src/fuzzer/generators/RandomInputGenerator.test.ts
@@ -1,8 +1,9 @@
+import * as JSON5 from "json5";
 import { ArgDef } from "../analysis/typescript/ArgDef";
-import { GeneratorFactory } from "./GeneratorFactory";
-import seedrandom from "seedrandom";
+import { RandomInputGenerator } from "./RandomInputGenerator";
 import { ProgramDef } from "../analysis/typescript/ProgramDef";
 import { ArgOptions, ArgValueType } from "../analysis/typescript/Types";
+import { FuzzIoElement } from "fuzzer/Types";
 
 /**
  * Provide a seed to ensure tests are deterministic.
@@ -119,13 +120,12 @@ describe("fuzzer/generator/GeneratorFactory:", () => {
  * @param intMax Maximim integer value
  */
 const testRandomInt = (intMin: number, intMax: number): void => {
-  const prng = seedrandom(seed);
   const program = ProgramDef.fromSource(() => tsFnWithNumberInput);
   const arg = program.getFunctions()["test"].getArgDefs();
   arg[0].setIntervals([{ min: intMin, max: intMax }]);
-  const gen = GeneratorFactory(arg[0], prng);
+  const gen = new RandomInputGenerator(arg, seed);
   for (let i = 0; i < 1000; i++) {
-    const result: ArgValueType = gen();
+    const result: ArgValueType = gen.next()[0].value;
     expect(typeof result === "number" && Number.isInteger(result)).toBeTruthy();
     expect(result).toBeGreaterThanOrEqual(intMin);
     expect(result).toBeLessThanOrEqual(intMax);
@@ -151,16 +151,15 @@ const testRandomIntException = (intMin: number, intMax: number): void => {
  * @param floatMax Maximim float value
  */
 const testRandomFloat = (floatMin: number, floatMax: number): void => {
-  const prng = seedrandom(seed);
   const program = ProgramDef.fromSource(() => tsFnWithNumberInput);
   const arg = program.getFunctions()["test"].getArgDefs();
   arg[0].setIntervals([{ min: floatMin, max: floatMax }]);
-  const gen = GeneratorFactory(arg[0], prng);
+  const gen = new RandomInputGenerator(arg, seed);
   for (let i = 0; i < 1000; i++) {
-    const result: ArgValueType = gen();
-    expect(typeof result === "number").toBeTruthy();
-    expect(result).toBeGreaterThanOrEqual(floatMin);
-    expect(result).toBeLessThanOrEqual(floatMax);
+    const input: ArgValueType = gen.next()[0].value;
+    expect(typeof input === "number").toBeTruthy();
+    expect(input).toBeGreaterThanOrEqual(floatMin);
+    expect(input).toBeLessThanOrEqual(floatMax);
   }
 };
 
@@ -184,20 +183,19 @@ const testRandomFloatException = (floatMin: number, floatMax: number): void => {
  * @param boolMax Maximum boolean value
  */
 const testRandomBool = (boolMin: boolean, boolMax: boolean): void => {
-  const prng = seedrandom(seed);
   const program = ProgramDef.fromSource(() => tsFnWithBoolInput);
   const arg = program.getFunctions()["test"].getArgDefs();
   arg[0].setIntervals([{ min: boolMin, max: boolMax }]);
-  const gen = GeneratorFactory(arg[0], prng);
+  const gen = new RandomInputGenerator(arg, seed);
 
   // Test that the generator generates booleans within the bounds
-  const results: ArgValueType[] = [];
+  const inputs: ArgValueType[] = [];
   for (let i = 0; i < 1000; i++) {
-    results.push(gen());
+    inputs[i] = gen.next()[0].value;
   }
-  expect(results.every((e) => typeof e === "boolean")).toBeTruthy();
-  expect(results.some((e) => e === boolMin)).toBeTruthy();
-  expect(results.some((e) => e === boolMax)).toBeTruthy();
+  expect(inputs.every((e) => typeof e === "boolean")).toBeTruthy();
+  expect(inputs.some((e) => e === boolMin)).toBeTruthy();
+  expect(inputs.some((e) => e === boolMax)).toBeTruthy();
 };
 
 /**
@@ -215,7 +213,6 @@ const testRandomString = (
   strLenMin: number,
   strLenMax: number
 ): void => {
-  const prng = seedrandom(seed);
   const options: ArgOptions = {
     ...ArgDef.getDefaultOptions(),
     strLength: { min: strLenMin, max: strLenMax },
@@ -228,26 +225,26 @@ const testRandomString = (
   const program = ProgramDef.fromSource(() => tsFnWithStringInput, options);
   const arg = program.getFunctions()["test"].getArgDefs();
   arg[0].setIntervals([{ min: strMin, max: strMax }]);
-  const gen = GeneratorFactory(arg[0], prng);
+  const gen = new RandomInputGenerator(arg, seed);
 
   // Test that the generator generates strings within the bounds
-  const results: ArgValueType[] = [];
+  const inputs: ArgValueType[] = [];
   for (let i = 0; i < 1000; i++) {
-    results.push(gen());
-  }
-  //if (strMin !== strMax) {
-  //  expect(results.some((e) => e !== strMin)).toBeTruthy();
-  //  expect(results.some((e) => e !== strMax)).toBeTruthy();
-  //}
-  results.forEach((result) => {
-    expect(typeof result === "string").toBeTruthy();
-    if (typeof result === "string") {
-      expect(result.length).toBeGreaterThanOrEqual(strLenMin);
-      expect(result.length).toBeLessThanOrEqual(strLenMax);
-      //expect(result >= strMin).toBeTruthy();
-      //expect(result <= strMax).toBeTruthy();
+    const input = gen.next()[0].value;
+    inputs[i] = input;
+    expect(typeof input === "string").toBeTruthy();
+    if (typeof input === "string") {
+      expect(input.length).toBeGreaterThanOrEqual(strLenMin);
+      expect(input.length).toBeLessThanOrEqual(strLenMax);
+
+      // expect(input >= strMin).toBeTruthy(); !!!! strMin/Max support
+      // expect(input <= strMax).toBeTruthy(); !!!! strMin/Max support
     }
-  });
+  }
+  if (strMin !== strMax) {
+    // expect(inputs.some((e) => e !== strMin)).toBeTruthy(); !!!! strMin/Max support
+    // expect(inputs.some((e) => e !== strMax)).toBeTruthy(); !!!! strMin/Max support
+  }
 };
 
 /**
