+++ src/fuzzer/measures/AbstractMeasure.ts
@@ -1,14 +1,30 @@
 import { FuzzTestResults } from "../Fuzzer";
 import { VmGlobals } from "../Types";
 
+// !!!!!!
 export abstract class AbstractMeasure {
   protected _tick = 0;
+  protected _weight = 1;
 
   /**
-   * Returns the measure's unique identifier
+   * Returns the measure's name
    */
-  public get id(): string {
+  public get name(): string {
     return this.constructor.name;
+  }
+
+  /**
+   * Returns the measure's weight
+   */
+  public get weight(): number {
+    return this._weight;
+  }
+
+  /**
+   * Set the measure's weight
+   */
+  public set weight(inWeight: number) {
+    this._weight = inWeight;
   }
 
   // !!!!!!
