+++ src/fuzzer/measures/MeasureFactory.ts
@@ -3,6 +3,6 @@ import { AbstractMeasure } from "./AbstractMeasure";
 import { CoverageMeasure } from "./CoverageMeasure";
 
 export function MeasureFactory(env: FuzzEnv): AbstractMeasure[] {
-  env; // !!!!!!! Base list of measures on FuzzEnv
+  env; // !!!!!!! Base list of measures on FuzzEnv; measure weight
   return [new CoverageMeasure()];
 }
